requirements.txt
python-dotenv
pyjwt
selenium
requests
